requirements.txt
pymerkle
kafka-python